Use a buffer for the comment contents
This isn't quite perfect. We kind of want to use botright split, and
jump manually back to the original window, but this seems to be
impossible to do on a WinClosed or similar event because the only time
vim pays attention to chaging the current buffer is in WinLeave, which
happens earlier (and before any BufHidden or BufWipeout). So we just
split the window with the actual code in it.

We use a temporrary file because that's actually kind of intuitive. You
:wq the file when you're done. In theory (if we had a more structured
temporrary file name) we could even recover previous comments lost after
a crash etc.

We rely on more vimspector code for a little UI popup for sasking about
raising an issue, but TBH i don't love this.